Golden Tier/.claude/skills/ceo-briefing/requirements.txt
xmlrpc>=1.3.2
python-dotenv>=1.0.0
